clearcut_detection_backend/model2/requirements.txt
catalyst==19.5
Flask==1.1.1
geopandas==0.5.1
google-api-python-client==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
imageio==2.5.0
rasterio==1.0.24
segmentation-models-pytorch==0.1.0
Shapely==1.6.4.post2
torch==1.4.0
torchvision==0.5.0
tqdm==4.19.9
oauth2client==4.1.3
Pillow==6.2.2
PyYAML==5.1.1

celery==4.4.6
SQLAlchemy==1.3.18
psycopg2-binary==2.8.5
